a bunch of fixes
