+++ lib/OrdinaryDiffEqExplicitRK/src/OrdinaryDiffEqExplicitRK.jl
@@ -1,7 +1,8 @@
 module OrdinaryDiffEqExplicitRK
 
 import OrdinaryDiffEq: alg_order, alg_adaptive_order, alg_stability_size, OrdinaryDiffEqAdaptiveAlgorithm,
-                       @cache, alg_cache, OrdinaryDiffEqConstantCache, @unpack, unwrap_alg
+                       @cache, alg_cache, OrdinaryDiffEqConstantCache, @unpack, unwrap_alg, 
+                       OrdinaryDiffEqMutableCache, initialize!, perform_step!
 using TruncatedStacktraces, RecursiveArrayTools, FastBroadcast, MuladdMacro
 
 include("algorithms.jl")
