+++ src/alg_utils.jl
@@ -36,6 +36,8 @@ dt_required(alg::LinearExponential) = false
 
 isdiscretealg(alg) = false
 
+function alg_stability_size end
+
 # evaluates f(t[i])
 _eval_index(f::F, t::Tuple{A}, _) where {F, A} = f(t[1])
 function _eval_index(f::F, t::Tuple{A, Vararg}, i) where {F, A}
