+++ src/doc_utils.jl
@@ -54,37 +54,6 @@ function generic_solver_docstring(description::String,
            "## References\n" * references
 end
 
-function rosenbrock_docstring(description::String,
-        name::String;
-        references::String = "",
-        extra_keyword_description = "",
-        extra_keyword_default = "",
-        with_step_limiter = false)
-    keyword_default = """
-    autodiff = Val{true}(),
-    concrete_jac = nothing,
-    linsolve = nothing,
-    precs = DEFAULT_PRECS,
-    """ * extra_keyword_default
-
-    keyword_default_description = """
-    - `autodiff`: boolean to control if the Jacobian should be computed via AD or not
-    - `concrete_jac`: function of the form `jac!(J, u, p, t)`
-    - `linsolve`: custom solver for the inner linear systems
-    - `precs`: custom preconditioner for the inner linear solver
-    """ * extra_keyword_description
-
-    if with_step_limiter
-        keyword_default *= "step_limiter! = OrdinaryDiffEq.trivial_limiter!,\n"
-        keyword_default_description *= "- `step_limiter!`: function of the form `limiter!(u, integrator, p, t)`\n"
-    end
-
-    generic_solver_docstring(
-        description, name, "Rosenbrock Method. ", references,
-        keyword_default_description, keyword_default
-    )
-end
-
 function explicit_rk_docstring(description::String,
         name::String;
         references::String = "",
@@ -105,36 +74,3 @@ function explicit_rk_docstring(description::String,
         keyword_default_description, keyword_default
     )
 end
-
-function rosenbrock_wanner_docstring(description::String,
-        name::String;
-        references::String = "",
-        extra_keyword_description = "",
-        extra_keyword_default = "",
-        with_step_limiter = false)
-    keyword_default = """
-    autodiff = Val{true}(),
-    concrete_jac = nothing,
-    linsolve = nothing,
-    precs = DEFAULT_PRECS,
-    """ * extra_keyword_default
-
-    keyword_default_description = """
-    - `autodiff`: boolean to control if the Jacobian should be computed via AD or not
-    - `concrete_jac`: function of the form `jac!(J, u, p, t)`
-    - `linsolve`: custom solver for the inner linear systems
-    - `precs`: custom preconditioner for the inner linear solver
-    """ * extra_keyword_description
-
-    if with_step_limiter
-        keyword_default *= "step_limiter! = OrdinaryDiffEq.trivial_limiter!,\n"
-        keyword_default_description *= "- `step_limiter!`: function of the form `limiter!(u, integrator, p, t)`\n"
-    end
-
-    generic_solver_docstring(
-        description, name, "Rosenbrock-Wanner Method. ", references,
-        keyword_default_description, keyword_default
-    )
-end
-
-# TODO other classes
